backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
python-dotenv==1.0.0